Make accordion into controlled components
Also divided sidebar tabs content into separate components
